+++ src/Symfony/Component/Config/Definition/EnumNode.php
@@ -55,4 +55,12 @@ class EnumNode extends ScalarNode
 
         return $value;
     }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function allowPlaceholders(): bool
+    {
+        return false;
+    }
 }